+++ test/resources_nav_test.rb
@@ -5,7 +5,39 @@ class ResourcesNavTest < ActiveSupport::TestCase
     assert ResourcesNav::VERSION
   end
 
-  test "it adds resources to an array" do
-    assert_equal [:apples, :oranges], Rails.application.routes.resources_nav
+  test "it adds resources to an array when nav option is present and true" do
+    [:apples, :oranges].each do |resource|
+      assert Rails.application.routes.resources_nav.include?(resource)
+    end
+  end
+
+  test "it doesn't add duplicates" do
+    Rails.application.routes.draw do
+      resources :apples
+    end
+    assert_equal Rails.application.routes.resources_nav.uniq, Rails.application.routes.resources_nav
+  end
+
+  test "it works also when multiple resources are specified" do
+    Rails.application.routes.draw do
+      resources :pears, :apricots, nav: true
+    end
+    [:apples, :oranges, :pears, :apricots].each do |resource|
+      assert Rails.application.routes.resources_nav.include?(resource)
+    end
+  end
+
+  test "it doesn't add resources to the array when nav option is false" do
+    Rails.application.routes.draw do
+      resources :sauerkrauts, nav: false
+    end
+    assert !Rails.application.routes.resources_nav.include?(:sauerkrauts)
+  end
+
+  test "it doesn't add resources to the array when nav option is not present" do
+    Rails.application.routes.draw do
+      resources :watermelons
+    end
+    assert !Rails.application.routes.resources_nav.include?(:watermelons)
   end
 end